backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# PDF processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
PyPDF2==3.0.1

# OCR
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# AI/ML
openai==1.3.7
anthropic==0.7.8

# Data processing
pandas==2.1.3
numpy==1.25.2

# Configuration
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
